Liquid: Another double -> float. MapHeaders: Added note, that some implicit conversion is most likely resulting in an error. Matrix: Now using the float variants of cos and sin. UIMenuBar: Now casting ints to floats explicitly. MenuItem: Now using a better constructor of UIButton to not need setText afterwards doing stuff twice.

+++ src/Matrix.h
@@ -115,8 +115,8 @@ public:
 
     //Rotate around the Y axis by B-90 
     rot.unit();
-    cosV=cos((r.y-90)*PI/180.0f);
-    sinV=sin((r.y-90)*PI/180.0f);
+    cosV=cosf((r.y-90.0f)*PI/180.0f);
+    sinV=sinf((r.y-90.0f)*PI/180.0f);
     rot.unit();
     rot.m[0][0]=cosV;
     rot.m[0][2]=sinV;
@@ -126,8 +126,8 @@ public:
     
     //Rotate around the Z axis by -A 
     rot.unit();
-    cosV=cos(-r.x*PI/180.0f);
-    sinV=sin(-r.x*PI/180.0f);
+    cosV=cosf(-r.x*PI/180.0f);
+    sinV=sinf(-r.x*PI/180.0f);
     rot.unit();
     rot.m[0][0]=cosV;
     rot.m[0][1]=sinV;
@@ -136,8 +136,8 @@ public:
     this->operator*=(rot);
 
     //Rotate around the X axis by C
-    cosV=cos(r.z*PI/180.0f);
-    sinV=sin(r.z*PI/180.0f);
+    cosV=cosf(r.z*PI/180.0f);
+    sinV=sinf(r.z*PI/180.0f);
     rot.unit();
     rot.m[1][1]=cosV;
     rot.m[1][2]=sinV;
